refactor: Moved ConsoleEmailValidatorRunner, changed its interface
