+++ app/src/Domain/ValueObjects/Email/EmailValidatorInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Pryaniki\App\Domain\ValueObjects\Email;
+
+use Pryaniki\App\Domain\ValueObjects\Email\ValidationResult;
+
+interface EmailValidatorInterface
+{
+    public function validate(string $email): ValidationResult;
+}
